WorkoutBot/requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.0
packaging==23.2
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==0.10.3
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.2